requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.2=py37heb32a55_1
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py37h06a4308_0
cycler=0.11.0=pyhd8ed1ab_0
freetype=2.10.4=h0708190_1
icu=67.1=he1b5a44_0
importlib-metadata=4.8.2=py37h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pyhd3eb1b0_0
kiwisolver=1.3.1=py37h2531618_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libllvm11=11.1.0=hf817b99_2
libpng=1.6.37=h21135ba_2
libstdcxx-ng=9.3.0=hd4cf53a_17
llvmlite=0.37.0=py37h295c915_1
matplotlib-base=3.2.2=py37h1d35a4c_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.3=h7f8727e_2
numba=0.54.1=py37h51133e4_0
numexpr=2.8.1=py37h6abb31d_0
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h8c16a72_0
patsy=0.5.2=pyhd8ed1ab_0
pip=21.2.2=py37h06a4308_0
pluggy=1.0.0=py37h06a4308_0
progressbar2=3.37.1=py37h06a4308_0
py=1.10.0=pyhd3eb1b0_0
pyod=0.9.7=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pytest=6.2.5=py37h06a4308_2
pytest-runner=5.3.1=pyhd3eb1b0_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-utils=2.5.6=py37h06a4308_0
python_abi=3.7=2_cp37m
pytz=2021.3=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
scikit-learn=1.0.1=py37h51133e4_0
scipy=1.7.3=py37hc147768_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
statsmodels=0.12.1=py37h902c9e0_2
tbb=2021.4.0=hd09550d_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py37h5e8e339_1
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
